restructured dashboard componenet to align clean code principles
