+++ src/App.js
@@ -2,9 +2,9 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import './App.css';
-import Dashboard from './Components/Dashboard';
-import Sidepanel from './Components/Sidepanel';
-import BlockedWebsites from './Components/BlockedWebsites';
+import Dashboard from './Components/Dashboard/Dashboard';
+import Sidepanel from './Components/Sidepanel/Sidepanel';
+import BlockedWebsites from './Components/BlockedWebsites/BlockedWebsites';
 
 function App() {
   return (
